requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
google-auth>=2.35
resend>=2.0
psycopg2-binary>=2.9
